requirements.txt
# PDF Processing
pdf2image==1.16.3
PyMuPDF==1.23.8
pdfplumber==0.10.3

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
scikit-image==0.22.0

# OCR Engines
pytesseract==0.3.10
easyocr==1.7.1

# AI/ML
anthropic==0.18.1
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2

# Data Models & Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Math Processing
sympy==1.12
latex2sympy2==1.9.1

# Utilities
tqdm==4.66.1
colorama==0.4.6
loguru==0.7.2

# API (Future)
fastapi==0.109.0
uvicorn==0.27.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
